changed row to unsigned int

+++ transition.c
@@ -44,7 +44,7 @@ Image linearTransitionV(Image *imageA, Image *imageB, float ratio) {
 }
 
 void revoluteTransition(Image *out, Image *previous, Image *current, float ratio) {
-  int row, col;
+  unsigned int row, col;
   float angle;
   int halfWidth, halfHeight;
   int inverse;
@@ -62,8 +62,8 @@ void revoluteTransition(Image *out, Image *previous, Image *current, float ratio
   angle = PI / 2.0F * ratio;
   halfWidth = image->width / 2;
   halfHeight = image->height / 2;
-  for(row = 0;row < (int)out->height;row++) {
-    for(col = 0;col < (int)out->width;col++) {
+  for(row = 0;row < out->height;row++) {
+    for(col = 0;col < out->width;col++) {
       long index = row * image->width + col;
       int x, y;
       x = col - halfWidth;
@@ -90,7 +90,7 @@ void revoluteTransition(Image *out, Image *previous, Image *current, float ratio
 }
 
 void stripeHTransitionBase(Image *out, Image *previous, Image *current, int n, float ratio, int direction) {
-  int row, col;
+  unsigned int row, col;
   int stride;
   int inverse;
   Image *image;
@@ -105,8 +105,8 @@ void stripeHTransitionBase(Image *out, Image *previous, Image *current, int n, f
   }
   stride = ceilf((float)image->width / n);
   ratio = max(min(ratio, 1.0F), 0.0F);
-  for(row = 0;row < (int)out->height;row++) {
-    for(col = 0;col < (int)out->width;col++) {
+  for(row = 0;row < out->height;row++) {
+    for(col = 0;col < out->width;col++) {
       long index = row * image->width + col;
       if((direction ? stride - col % stride : col % stride) <= ratio * stride) {
         out->data[index] = inverse ? BLACK : image->data[index];
@@ -118,7 +118,7 @@ void stripeHTransitionBase(Image *out, Image *previous, Image *current, int n, f
 }
 
 void stripeVTransitionBase(Image *out, Image *previous, Image *current, int n, float ratio, int direction) {
-  int row, col;
+  unsigned int row, col;
   int stride;
   int inverse;
   Image *image;
@@ -133,8 +133,8 @@ void stripeVTransitionBase(Image *out, Image *previous, Image *current, int n, f
   }
   stride = ceilf((float)image->height / n);
   ratio = max(min(ratio, 1.0F), 0.0F);
-  for(row = 0;row < (int)out->height;row++) {
-    for(col = 0;col < (int)out->width;col++) {
+  for(row = 0;row < out->height;row++) {
+    for(col = 0;col < out->width;col++) {
       long index = row * image->width + col;
       if((direction ? stride - row % stride : row % stride) <= ratio * stride) {
         out->data[index] = inverse ? BLACK : image->data[index];
